[android] #6042 - bring isConnected() to MapboxAccountManager and make its private SDK use more explicit
