+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/MapboxAccountManager.java
@@ -4,12 +4,12 @@ import android.content.Context;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.text.TextUtils;
+import android.util.Log;
 
 import com.mapbox.mapboxsdk.constants.MapboxConstants;
 import com.mapbox.mapboxsdk.exceptions.InvalidAccessTokenException;
 import com.mapbox.mapboxsdk.exceptions.MapboxAccountManagerNotStartedException;
 import com.mapbox.mapboxsdk.telemetry.MapboxEventManager;
-import com.mapbox.mapboxsdk.utils.IOUtils;
 
 public class MapboxAccountManager {
 
@@ -43,7 +43,7 @@ public class MapboxAccountManager {
         if (mapboxAccountManager == null) {
             mapboxAccountManager = new MapboxAccountManager(context, accessToken);
         }
-        IOUtils.getInstance(mapboxAccountManager.applicationContext);
+
         MapboxEventManager eventManager = MapboxEventManager.getMapboxEventManager();
         eventManager.initialize(mapboxAccountManager.applicationContext, mapboxAccountManager.accessToken);
         return mapboxAccountManager;
@@ -83,4 +83,19 @@ public class MapboxAccountManager {
             throw new InvalidAccessTokenException();
         }
     }
+
+    /**
+     * Determines whether we have an Internet connection available. Please do not rely on this
+     * method in your apps, this method is used internally by the SDK.
+     *
+     * @return true if there is an Internet connection, false otherwise
+     */
+    public boolean isConnected() {
+        ConnectivityManager cm = (ConnectivityManager) applicationContext.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
+        boolean result = (activeNetwork != null && activeNetwork.isConnected());
+        Log.i("IOUtils", "isConnected result = " + result);
+        return result;
+    }
+
 }
